quality_bot/requirements.txt
python-telegram-bot>=22.0
requests
feedparser
newspaper3k
beautifulsoup4
readability-lxml
arabic-reshaper
python-bidi
nltk
reportlab
jinja2
weasyprint
pillow
lxml
boto3>=1.34.0
python-dotenv>=1.0.0
playwright>=1.40.0
apify-client>=1.7.0
